Test on PHP 7.2 & 7.4
Before: the CI ran directly on Ubuntu 20.04, so would just use the PHP version that provides (7.4).

Now: we're using the `shivammathur/setup-php@v2` to directly control what PHP versions we're running CI on. That means we can test on both 7.2 & 7.4, and will make it easier to cater for other versions in future.