Downgrade theme to PHP 7.2
Before: theme required at least 7.4. This will be fine for anything running on Dalmatian or dxw GovPress infrastructure, but legacy Helpful infrastructure is still on PHP 7.2, and it's possible we'll be creating new themes that will be hosted there, at least in the immediate future.

Now: the min PHP version for the theme is now 7.2