Add build artifacts section
These are based on the rules that would previously have lived in the
theme-specific .gitignore in wp-content/themes/theme/.

However, that file also includes some rules we'll be duplicating at the
top-level (e.g. ignoring vendor/ folders), and some rules I'm not sure
we should be applying by default (e.g. ignoring .map files). So I've
removed the ones we'll be duplicating, or don't need.