Add sensitive files section to .gitignore
This is lifted wholesale from the dxw Rails template. As a result, there
are a number of patterns covered here (e.g. the Terraform-related ones)
that are highly unlikely to occur in a GovPress project. However, given
that the intention of this section is to ensure that sensitive data is
not committed accidentally, the additional safety of including those
rules feels like it outweighs the readability/maintenance cost.