Remove composer.lock from project root
The composer.json only contains a single dev dependency: Whippet.

We're not going to be installing and running this anywhere except on our
own machines or in CI.

Therefore, there is no benefit to locking Whippet to a specific version,
as it prevents us from automatically picking up non-breaking updates
without manually updating the composer.lock file. So let's just remove
the composer.lock.