Suppress non-security PRs from Dependabot
The current dependabot.yml rules will result in Dependabot opening PRs
for any composer package updates available, or any prod NPM package
updates available.

In theory, this is a desirable state of affairs, but because of our
current build process (where we wrap composer dependencies up into a
vendor.phar file we commit, and we run asset compilation locally and
commit the result) these PRs would require manual intervention to have a
useful output. That isn't sustainable at the moment.

The long-term solution here is probably to have a more standardised
build process, where dependency installation and asset compilation
happen as part of the automated build in CI. But for now, let's just
switch off those PRs so we don't get overwhelmed.

We don't need to switch off PRs for the root composer file, because this
just contains Whippet as a dev dependency. We very rarely give Whippet a
major version bump, but when we do we'll probably want this reflected
in each project using it.