Allow customisation of header logo's alt text
As part of settle's accessibility review, we noticed that the logo alt
text uses to the WordPress site description, which defaults to "just
another wordpress site".

This changes the default value to the site's tite instead. This is
overridable by setting the "header logo alt text" in the "Global Content
and Options" section of the admin site. This section has been used
elsewhere, so it made sense to make it part of the template too.

NB: The current template doesn't include a logo image by default, so
I've provided a commented-out example in `global-header.php` and
modified the logo's `<h1>` tag to display the alt text.