Add workflow for running GhostInspector tests on prod deploy PR
This workflow will run on PRs merging `develop` into `main` (i.e.
production deploy PRs). It will use the secrets set automatically by
GovPress Tools to run the current project's staging Ghost Inspector test
suite, and fail if that is not passing (or if the secrets are not set,
i.e. the test suite has not been created).