Allow script/setup to be run before script/server
The expected convention when using the "scripts to rule them all"
pattern is that you run `script/setup` in order to get your project into
a known state, and then `script/server` to run a local server to access
it.

However, traditionally we've required `script/server` to be run first,
because the server has to be running in order for WordPress to be
installed (and for internal WordPress config to be successful).

This change updates the setup script so it starts the server, if it's
not already running, then wait to confirm that the database container is
ready for user (by pinging it from the WordPress container), and then
runs the setup script. It then stops the server again, so that the
server is only running if the user has chosen to run `script/server`.