Reset database, with option to backup, on setup
Before: when you ran the setup script, it wouldn't clear out the
database before running, it would just run the setup steps on top of
your existing database state. This was confusing, and not really what
the setup script is intended for (setting the site back to a known
state).

Now: when you run the setup script, the database will be reset, and the
setup steps run from a blank slate. In order to avoid this
unintentionally over-writing people's current database state, it checks
if a WordPress database currently exists, and if so prompts if you'd
like to take a backup. If you do, the script then provides you with the
command you'll need to run to restore that backup at a later date.