Run `composer update` rather than `composer install` for root deps
Because we don't commit a `composer.lock` file for the root of the
project, it's possible that we could have modified the root
`composer.json` and the changes wouldn't get picked up by another
`composer install` of the user had already run it locally (as that would
just install the deps already in their local `composer.lock`).

So let's run `composer update` each time.